+++ src/OpenTelemetry.Instrumentation.Http/HttpClientInstrumentationOptions.cs
@@ -18,7 +18,6 @@ using System;
 using System.Diagnostics;
 using System.Net.Http;
 using System.Runtime.CompilerServices;
-using OpenTelemetry.Context.Propagation;
 using OpenTelemetry.Instrumentation.Http.Implementation;
 using OpenTelemetry.Trace;
 
@@ -33,12 +32,6 @@ namespace OpenTelemetry.Instrumentation.Http
         /// Gets or sets a value indicating whether or not the HTTP version should be added as the <see cref="SemanticConventions.AttributeHttpFlavor"/> tag. Default value: False.
         /// </summary>
         public bool SetHttpFlavor { get; set; }
-
-        /// <summary>
-        /// Gets or sets <see cref="TextMapPropagator"/> for context propagation.
-        /// By default, <see cref="Propagators.DefaultTextMapPropagator" /> will be used.
-        /// </summary>
-        public TextMapPropagator Propagator { get; set; } = Propagators.DefaultTextMapPropagator;
 
         /// <summary>
         /// Gets or sets a Filter function to filter instrumentation for requests on a per request basis.