+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -6,8 +6,11 @@
   to CompositeTextMapPropagator. IPropagator is renamed to TextMapPropagator
   and changed from interface to abstract class.
   ([#1427](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1427))
-* Propagators.DefaultTextMapPropagator will be used as the default Propagator
+* Propagators.DefaultTextMapPropagator will be used as the default Propagator.
   ([#1427](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1428))
+* Removed Propagator from Instrumentation Options. Instrumentation now always
+  respect the Propagator.DefaultTextMapPropagator.
+  ([#1448](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1448))
 
 ## 0.7.0-beta.1